requirements.txt
streamlit>=1.36
pandas>=2.1
numpy>=1.26
biopython>=1.83
plotly>=5.22
openpyxl>=3.1
mygene>=3.2
matplotlib>=3.8
scikit-learn 
shap
umap-
tomli>=2.0.1
#ViennaRNA>=2.6       # Python bindings; optional but nice to have